+++ src/main.ts
@@ -1,9 +1,20 @@
-﻿export * from './model/character.interface';
+﻿export * from './model/asset/asset.entity';
+export * from './model/asset/asset-instance.entity';
+export * from './model/asset/asset-type.entity';
+export * from './model/asset/body-part.enum';
+export * from './model/asset/equip-effect.entity';
+export * from './model/asset/equip-effect.enum';
 
-export * from './model/asset.interface';
-export * from './model/effect.interface';
-export * from './model/place.interface';
-export * from './model/story.interface';
+export * from './model/attribute/attribute.entity';
 
-export * from './controller/character.controller';
-export * from './controller/game.controller';
+export * from './model/character/character-equipment.entity';
+export * from './model/character/character-map.entity';
+export * from './model/character/character.interface';
+
+export * from './model/place/place.entity';
+export * from './model/place/location.entity';
+
+export * from './model/story/asset-condition.entity';
+export * from './model/story/attribute-condition.entity';
+export * from './model/story/decision.entity';
+export * from './model/story/story.entity';