+++ src/model/asset/body-part.enum.ts
@@ -1,0 +1,8 @@
+﻿/** all body parts that can have equipment */
+export enum BodyPart {
+	Head = 'Head',
+	Body = 'Body',
+	Wrist = 'Wrist',
+	Necklace = 'Necklace',
+	Shoes = 'Shoes',
+}