+++ src/model/asset/equip-effect.entity.ts
@@ -1,0 +1,8 @@
+﻿import { Entity } from '../entity.interface';
+import { Attribute } from '../attribute/attribute.entity';
+import { EquipEffectType } from './equip-effect.enum';
+
+export interface EquipEffect extends Entity {
+	attribute: Attribute;
+	type: EquipEffectType;
+}