+++ src/model/character/character-equipment.entity.ts
@@ -1,0 +1,9 @@
+﻿import { EquippableAsset } from '../asset/asset.entity';
+
+export interface CharacterEquipment {
+	head?: EquippableAsset;
+	body?: EquippableAsset;
+	wrist?: EquippableAsset;
+	necklace?: EquippableAsset;
+	shoes?: EquippableAsset;
+}