+++ src/model/story/attribute-condition.entity.ts
@@ -1,0 +1,6 @@
+﻿import { Attribute } from '../attribute/attribute.entity';
+
+export interface AttributeCondition {
+	attributeToActivate: Attribute;
+	attributeLevelFor100Percent: number;
+}