+++ src/test/data/attributes.object.ts
@@ -1,9 +1,9 @@
-﻿import { AssetType, Attribute } from '../../model/asset.interface';
+﻿import { Attribute } from '../../model/attribute/attribute.entity';
 
-export const attributes: { [key: string]: Attribute } = {
-	Strength: {
+export const Attributes: Attribute[] = [
+	{
+		uuid: '8a098f03-2e0a-486a-b7f4-151c9fefbcb2',
 		name: 'Strength',
-		type: AssetType.Attribute,
 		pointsCollected: 1,
 	},
-};
+];