+++ src/test/game.spec.ts
@@ -1,5 +1,4 @@
-﻿import { CharacterObject } from './data/character.object';
-import { StoryCafeObject } from './data/story-cafe.object';
+﻿import { CharacterObject, StoryCafeObject } from './data/story-cafe.object';
 import { GameController } from '../controller/game.controller';
 
 describe('game simulation', () => {
@@ -13,8 +12,8 @@ describe('game simulation', () => {
 			stories,
 		});
 		expect(decisions.length).toBe(2);
-		expect(decisions[0].title).toBe('Drink some coffee');
-		expect(decisions[1].title).toBe('Work: Wash dishes');
+		expect(decisions[0].name).toBe('Drink some coffee');
+		expect(decisions[1].name).toBe('Work: Wash dishes');
 
 		// They want coffee? They'll get coffee, let them drink that stuff
 		const result = GameController.submitDecision({
@@ -24,11 +23,30 @@ describe('game simulation', () => {
 		});
 		expect(result.characterWins).toBeTruthy();
 		expect(result.text).toBe('Ah, refreshing as always.');
-		// the character should be updated accordingly
-		const coffee = character.assets.get('Coffee');
-		expect(coffee).toStrictEqual([{ name: 'Coffee', type: 'Normal' }, 1]);
-		const money = character.assets.get('Money');
-		expect(money).toStrictEqual([{ name: 'Money', type: 'Normal' }, 1]);
+		// the character should be updated accordingly and retrieve coffee
+		const coffee = character.assetInstances.find(
+			(i) => i.asset.uuid === 'abcd213b-2691-45ae-8d51-9a1d0f0ab805'
+		);
+		expect(coffee).toStrictEqual({
+			asset: {
+				name: 'Coffee',
+				type: 'Normal',
+				uuid: 'abcd213b-2691-45ae-8d51-9a1d0f0ab805',
+			},
+			count: 1,
+		});
+		// Player lost some money
+		const money = character.assetInstances.find(
+			(i) => i.asset.uuid === '2b1793f7-6349-4bfc-b12e-ad25eba5a485'
+		);
+		expect(money).toStrictEqual({
+			asset: {
+				name: 'Money',
+				type: 'Normal',
+				uuid: '2b1793f7-6349-4bfc-b12e-ad25eba5a485',
+			},
+			count: 1,
+		});
 
 		// Another coffee? This should throw due to the lack of money
 		const sd1 = () =>
@@ -37,6 +55,7 @@ describe('game simulation', () => {
 				character: character,
 				stories: stories,
 			});
+		console.log(character.assetInstances);
 		expect(sd1).toThrow('INVALID_DECISION');
 
 		// seems like you should work for your coffee
@@ -46,11 +65,31 @@ describe('game simulation', () => {
 			stories: stories,
 		});
 		expect(result2.characterWins).toBeTruthy();
-		const assets = Object.fromEntries(character.assets);
-		expect(assets).toStrictEqual({
-			Money: [{ name: 'Money', type: 'Normal' }, 4],
-			Coffee: [{ name: 'Coffee', type: 'Normal' }, 1],
-			'Job: Café': [{ name: 'Job: Café', type: 'Normal' }, 1],
-		});
+		expect(character.assetInstances).toStrictEqual([
+			{
+				asset: {
+					name: 'Money',
+					type: 'Normal',
+					uuid: '2b1793f7-6349-4bfc-b12e-ad25eba5a485',
+				},
+				count: 4,
+			},
+			{
+				asset: {
+					name: 'Coffee',
+					type: 'Normal',
+					uuid: 'abcd213b-2691-45ae-8d51-9a1d0f0ab805',
+				},
+				count: 1,
+			},
+			{
+				asset: {
+					name: 'Job: Café',
+					type: 'Normal',
+					uuid: 'eaca665a-cb7a-491b-b14d-b279754bebf0',
+				},
+				count: 1,
+			},
+		]);
 	});
 });